Refactor, namespacing, comments, etc.
