+++ src/AgentFramework.AspNetCore/AgentBuilderExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using AgentFramework.AspNetCore.Hosting;
 using AgentFramework.Core.Contracts;
 using AgentFramework.Core.Handlers;
 using AgentFramework.Core.Handlers.Agents;
