+++ src/AgentFramework.AspNetCore/Hosting/AgentHostedService.cs
@@ -9,8 +9,9 @@ using Hyperledger.Indy.WalletApi;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 
-namespace AgentFramework.AspNetCore
+namespace AgentFramework.AspNetCore.Hosting
 {
+    /// <inheritdoc />
     /// <summary>
     /// Agent hosted service.
     /// </summary>
@@ -21,7 +22,7 @@ namespace AgentFramework.AspNetCore
         private readonly PoolOptions _poolOptions;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="T:AgentFramework.AspNetCore.AgentHostedService"/> class.
+        /// Initializes a new instance of the <see cref="T:AgentFramework.AspNetCore.Hosting.AgentHostedService"/> class.
         /// </summary>
         /// <param name="provisioningService">Provisioning service.</param>
         /// <param name="provisioningConfiguration">Provisioning configuration.</param>
@@ -45,6 +46,7 @@ namespace AgentFramework.AspNetCore
         /// <value>The provisioning configuration.</value>
         public ProvisioningConfiguration ProvisioningConfiguration { get; }
 
+        /// <inheritdoc />
         /// <summary>
         /// Triggered when the application host is ready to start the service.
         /// </summary>
@@ -81,6 +83,7 @@ namespace AgentFramework.AspNetCore
             }
         }
 
+        /// <inheritdoc />
         /// <summary>
         /// Triggered when the application host is performing a graceful shutdown.
         /// </summary>
