+++ src/AgentFramework.Core.Handlers/Internal/DefaultForwardHandler.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using AgentFramework.Core.Contracts;
 using AgentFramework.Core.Extensions;
+using AgentFramework.Core.Handlers.Agents;
 using AgentFramework.Core.Messages;
 using AgentFramework.Core.Messages.Routing;
