+++ src/AgentFramework.Core/Models/Payments/PaymentAmount.cs
@@ -2,11 +2,20 @@
 
 namespace AgentFramework.Core.Models.Payments
 {
+    /// <summary>
+    /// Payment amount
+    /// </summary>
     public class PaymentAmount
     {
+        /// <summary>
+        /// Gets or sets the currency
+        /// </summary>
         [JsonProperty("currency")]
         public string Currency { get; set; }
 
+        /// <summary>
+        /// Gets or sets the payment value
+        /// </summary>
         [JsonProperty("value")]
         public ulong Value { get; set; }
     }
