+++ src/AgentFramework.TestHarness/Mock/MockAgent.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using AgentFramework.Core.Contracts;
 using AgentFramework.Core.Handlers;
+using AgentFramework.Core.Handlers.Agents;
 using AgentFramework.Core.Messages;
 using Microsoft.Extensions.DependencyInjection;
