+++ test/AgentFramework.Core.Tests/Decorators/SignatorDecoratorTests.cs
@@ -5,6 +5,7 @@ using AgentFramework.Core.Decorators.Signature;
 using AgentFramework.Core.Decorators.Threading;
 using AgentFramework.Core.Exceptions;
 using AgentFramework.Core.Handlers;
+using AgentFramework.Core.Handlers.Agents;
 using AgentFramework.Core.Messages.Connections;
 using AgentFramework.Core.Models.Connections;
 using AgentFramework.Core.Utils;
